event-producer/requirements.txt
kafka-python==2.0.2
pytest==7.1.3
jsonref==1.0.0
typing-extensions==4.4.0
beautifulsoup4==4.11.1
APScheduler==3.9.1.post1
jsonschema==4.17.1
requests==2.28.1
pyyaml==6.0
mypy==0.991
overrides==7.3.1